+++ stretch_fetch_grasp_bridge/scripts/stretch_grasp_filter.py
@@ -5,68 +5,135 @@ from geometry_msgs.msg import PoseArray
 from geometry_msgs.msg import PoseStamped, Pose
 from std_msgs.msg import Int32, String
 import tf.transformations as tf_transform
+from stretch_fetch_grasp_bridge.srv import StretchGraspPosev2, StretchGraspPosev2Response
+from rail_manipulation_msgs.srv import SuggestGrasps, SuggestGraspsRequest, SuggestGraspsResponse
 import numpy as np
 
 
-class stretch_grasp_filter:
+class StretchGraspFilter:
     def __init__(self):
-        # configs
-        self.filter_angle_buffer = 10 # degrees
+        rospy.init_node('stretch_grasp_filter', anonymous=True)
         
-        rospy.init_node('pose_array_listener', anonymous=True)
+        
+        # configs
+        self.filter_angle_buffer = rospy.get_param('/stretch_grasp/filter_angle_buffer', 'all') #10 # (in degrees) the buffer around the filter_angle
 
-        # subscribers and publishers
-        self.all_grasp_sub = rospy.Subscriber('/test_grasp_suggestion/all_grasp_poses', PoseArray, self.pose_array_callback)
+        # params
+        self.type_of_grasps = rospy.get_param('/stretch_grasp/filtering_type', 'all')
+        # possible values:
+        ## "frontal": grasps with the gripper pointing parallel to the ground plane
+        ## "top_down": grasps with the gripper pointing directly down
+        ## "lateral": grasps that have the antipodal point in a line parallel to the ground plane.
+        ## "all": all grasps without any filtering
+        
+        
+        # Publishers
         self.filter_grasp_pub = rospy.Publisher('/stretch_grasp/filtered_grasp_poses', PoseArray, queue_size=10)
-        self.last_pos_req_sub = rospy.Subscriber('/stretch_grasp/last_pos_req',Int32, self.last_pos_req_sub_callback)
         self.single_grasp_pub = rospy.Publisher('/stretch_grasp/single_grasp_pose', PoseStamped, queue_size=10)
-        self.last_pose_array = None
-        self.recovery_pub = rospy.Publisher('/stretch_grasp/recovery', String, queue_size=10)
-        self.type_of_grasps = rospy.get_param('filtering_type', 'frontal') # possible values = 'all', 'frontal', 'top_down', 'lateral'
         self.debug_grasp_pub = rospy.Publisher('/stretch_grasp/debug_grasp_pose', PoseStamped, queue_size=10)
 
-        # frontal grasps are grasps with the gripper pointing parallel to the ground plane
-        # top_down grasps are grasps with the gripper pointing directly down
-        # lateral grasps are grasps that have the antipodal point in a line parallel to the ground plane.
-        # all grasps are all grasps without any filtering
+        # Services
+        ## server
+        self.stretch_grasp_service = rospy.Service('/stretch_grasp_pose_suggester', StretchGraspPosev2, self.stretch_grasp_service_callback)
+        rospy.loginfo("Waiting for fetch suggestion service")
+        rospy.wait_for_service('/suggester/suggest_grasps')
 
-        # self.stretch_grasp_pose_service = rospy.Service('/stretch_grasp/get_grasp_pose', PoseArray, self.pose_array_callback)
-    def debug_pose_print(self,pose):
-        quat = [pose.orientation.x, pose.orientation.y, pose.orientation.z, pose.orientation.w]
-        point_in_coord1 = self.point_in_coordinate1(quat)
-        print("Point in coord1 is: ", point_in_coord1)
-        ang = self.angle_between_vectors(point_in_coord1, [0,0,1])
-        deg_ang = np.rad2deg(ang)
-        rospy.loginfo("#"*10)
-        rospy.loginfo("Quaternion is: %f, %f, %f, %f", quat[0], quat[1], quat[2], quat[3])  
-        rospy.loginfo("Angle is: %f", deg_ang)
-        rospy.loginfo("Point in coord1 is: %f, %f, %f", point_in_coord1[0], point_in_coord1[1], point_in_coord1[2])
-        rospy.loginfo("Point in coord2 is: %f, %f, %f", 1, 0, 0)
+        ## client
+        self.fetch_grasp_client = rospy.ServiceProxy('/suggester/suggest_grasps', SuggestGrasps)
+        rospy.loginfo("Connected to fetch suggestion service")
         
-    def last_pos_req_sub_callback(self,data):
-        rospy.loginfo("Received last_pos_req: %d", data.data)
-        if self.last_pose_array is None:
-            rospy.loginfo("last_pose_array is None")
-            return
-        if data.data < len(self.last_pose_array.poses):
-            rospy.loginfo("Publishing single grasp pose")
-            single_pose = PoseStamped()
-            single_pose.header = self.last_pose_array.header
-            single_pose.pose = self.last_pose_array.poses[data.data]
-            # self.debug_pose_print(single_pose.pose)
-            aligned_pose = self.align_new_pose_frame(single_pose) # aligns the pose frame to the base_link frame's z axis and maintains the antipodal points.
-            # displaced_pose = self.displce_pose([0.035,0,0],aligned_pose)
-            displaced_pose = self.displce_pose([0.0,0,0],aligned_pose)
+    
+    def stretch_grasp_service_callback(self,request):
+        '''
+        Callback function for the stretch_grasp_pose_suggester service. 
+        This function takes in a point cloud and returns a PoseStamped message type containing the pose of the best grasp.
 
-            self.single_grasp_pub.publish(displaced_pose)
-            self.debug_grasp_pub.publish(aligned_pose)
-            
+        Parameters:
+        - request: A StretchGraspPosev2Request message type containing the point cloud.
+
+        Returns:
+        - A StretchGraspPosev2Response message type containing the pose of the best grasp.
+        '''
+        rospy.logdebug("Received grasp request")
+        
+        # Fetch Grasp
+        fetch_request = SuggestGraspsRequest()
+        fetch_request.cloud = request.point_cloud
+        fetch_grasp_resp = self.fetch_grasp_client(fetch_request)
+        grasp_list = fetch_grasp_resp.grasp_list
+
+        # Filter Grasp
+        filtered_grasps = self.filter_grasps(grasp_list)
+        selected_grasp = filtered_grasps.poses[0]
+        
+        # Transform Grasp for viable Stretch Grasp
+        aligned_grasp = self.align_new_pose_frame(selected_grasp)
+        displaced_grasp = self.displace_pose([0.03,0.0,0.0],aligned_grasp)
+
+
+        resp = StretchGraspPosev2Response()
+        resp.grasp_pose.header.frame_id = 'base_link'
+        resp.grasp_pose.header.stamp = rospy.Time.now()
+        
+        resp.grasp_pose.pose = aligned_grasp
+        self.debug_grasp_pub.publish(resp.grasp_pose)
+
+        resp.grasp_pose.pose = displaced_grasp
+        self.single_grasp_pub.publish(resp.grasp_pose)
+        resp.success = True
+        return resp
+
+    def filter_grasps(self,grasp_list)->PoseArray:
+        """
+        Filters the grasps based on the type_of_grasps parameter.
+
+        Parameters:
+        - grasp_list: A PoseArray message type containing the list of grasps to be filtered.
+
+        Returns:
+        - A PoseArray message type containing the filtered grasps.
+        """
+        self.type_of_grasps = rospy.get_param('filtering_type', 'all')
+        rospy.logdebug("type_of_grasps is: " + str(self.type_of_grasps))
+        rospy.logdebug("Received PoseArray with %d poses.", len(grasp_list.poses))
+        if(self.type_of_grasps == 'all'):
+            self.filter_grasp_pub.publish(grasp_list)
+            self.filtered_pose_array = grasp_list
+            rospy.logdebug("Published PoseArray with %d poses.", len(grasp_list.poses))
+            return(grasp_list)
+        if(self.type_of_grasps == 'frontal'):
+            self.filter_angle = 90 # degrees
+            axis_of_interest = [1,0,0]
+            self.filter_angle_buffer = 10
+        elif(self.type_of_grasps == 'top_down'):
+            self.filter_angle = 180
+            axis_of_interest = [1,0,0]
+            self.filter_angle_buffer = 20
+        elif(self.type_of_grasps == 'lateral'):
+            self.filter_angle = 90
+            axis_of_interest = [0,1,0]
+            self.filter_angle_buffer = 10
         else:
-            rospy.loginfo("last_pos_req is out of bounds")
-            self.recovery_pub.publish("last_pos_req is out of bounds")
-            return
+            rospy.logerr("Unknown type of grasps for filtering. Defaulting to all grasps")
+            self.filter_grasp_pub.publish(grasp_list)
+            self.filtered_pose_array = grasp_list
         
-
+        filtered_poses = PoseArray()
+        filtered_poses.header = grasp_list.header
+        
+        for i, pose in enumerate(grasp_list.poses):
+            quat = [pose.orientation.x, pose.orientation.y, pose.orientation.z, pose.orientation.w]
+            point_in_coord1 = self.point_in_coordinate1(quat,axis_of_interest) # returns the vector axis_of_interest vector in the base_link frame.
+            ang = self.angle_between_vectors(point_in_coord1, [0,0,1])
+            deg_ang = np.rad2deg(ang)
+          
+            if((deg_ang> self.filter_angle-self.filter_angle_buffer) and (deg_ang<self.filter_angle+self.filter_angle_buffer)): # can make this a parameter
+                filtered_poses.poses.append(pose)
+        self.filter_grasp_pub.publish(filtered_poses)
+        self.filtered_pose_array = filtered_poses 
+        rospy.logdebug("Published PoseArray with %d poses.", len(filtered_poses.poses))
+        return(filtered_poses)
+        
     def angle_between_vectors(self,A, B):
         """
         Returns the angle (in radians) between vectors A and B.
@@ -117,49 +184,7 @@ class stretch_grasp_filter:
         # point_in_coord1 = np.dot(inv_rotation_matrix, point_in_coord2)
 
         return point_in_coord1
-
-    def pose_array_callback(self,data):
-        self.type_of_grasps = rospy.get_param('filtering_type', 'frontal')
-        rospy.loginfo("type_of_grasps is: " + str(self.type_of_grasps))
-        rospy.loginfo("Received PoseArray with %d poses.", len(data.poses))
-        if(self.type_of_grasps == 'all'):
-            self.filter_grasp_pub.publish(data)
-            self.last_pose_array = data
-            rospy.loginfo("Published PoseArray with %d poses.", len(data.poses))
-            return
-        if(self.type_of_grasps == 'frontal'):
-            self.filter_angle = 90 # degrees
-            axis_of_interest = [1,0,0]
-            self.filter_angle_buffer = 10
-        elif(self.type_of_grasps == 'top_down'):
-            self.filter_angle = 180
-            axis_of_interest = [1,0,0]
-            self.filter_angle_buffer = 20
-        elif(self.type_of_grasps == 'lateral'):
-            self.filter_angle = 90
-            axis_of_interest = [0,1,0]
-            self.filter_angle_buffer = 10
-        else:
-            print("Unknown type of grasps.")
-        # self.filter_angle = 90 # degrees
-        filtered_poses = PoseArray()
-        filtered_poses.header = data.header
-        print("Header is: ", data.header)
-        for i, pose in enumerate(data.poses):
-            quat = [pose.orientation.x, pose.orientation.y, pose.orientation.z, pose.orientation.w]
-            point_in_coord1 = self.point_in_coordinate1(quat,axis_of_interest) # returns the vector axis_of_interest vector in the base_link frame.
-            ang = self.angle_between_vectors(point_in_coord1, [0,0,1])
-            deg_ang = np.rad2deg(ang)
-            print("##")
-            print('deg_ang:',deg_ang)
-            if((deg_ang> self.filter_angle-self.filter_angle_buffer) and (deg_ang<self.filter_angle+self.filter_angle_buffer)): # can make this a parameter
-                rospy.logdebug("Angle is: "+ str(deg_ang))
-                rospy.loginfo("Angle is: "+ str(deg_ang))
-                filtered_poses.poses.append(pose)
-        self.filter_grasp_pub.publish(filtered_poses)
-        self.last_pose_array = filtered_poses 
-        rospy.loginfo("Published PoseArray with %d poses.", len(filtered_poses.poses))
-
+    
     def quaternion_to_rotation_matrix(self,quaternion): # Credit to ChatGPT4
         """
         Converts a quaternion into a rotation matrix.
@@ -225,14 +250,13 @@ class stretch_grasp_filter:
         of the z-axis as the base_link frame's.
 
         Args:
-        pose_frame (dict): A dictionary containing 'position' and 'orientation' of the original pose.
-                        'orientation' should be a quaternion (x, y, z, w).
+        pose (Pose): A Pose message type.
 
         Returns:
-        dict: A new pose with the specified orientation.
+        Pose: A new pose which is aligned to the z-axis of the base_link frame and has the similar y-axis as the original pose.
         """
         # Extract the rotation matrix from the original pose's quaternion
-        quaternion = [pose.pose.orientation.x, pose.pose.orientation.y, pose.pose.orientation.z, pose.pose.orientation.w]
+        quaternion = [pose.orientation.x, pose.orientation.y, pose.orientation.z, pose.orientation.w]
         rotation_matrix = self.quaternion_to_rotation_matrix(quaternion)
 
         # Extract the y-axis from the original pose's rotation matrix
@@ -258,41 +282,40 @@ class stretch_grasp_filter:
         new_orientation = self.rotation_matrix_to_quaternion(new_rotation_matrix)
 
         # Create new pose
-        new_pose = PoseStamped()
-        new_pose.header = pose.header
-        new_pose.pose.position = pose.pose.position
-        new_pose.pose.orientation.x = new_orientation[0]
-        new_pose.pose.orientation.y = new_orientation[1]
-        new_pose.pose.orientation.z = new_orientation[2]
-        new_pose.pose.orientation.w = new_orientation[3]
+        new_pose = Pose()
+        
+        new_pose.position = pose.position
+        new_pose.orientation.x = new_orientation[0]
+        new_pose.orientation.y = new_orientation[1]
+        new_pose.orientation.z = new_orientation[2]
+        new_pose.orientation.w = new_orientation[3]
 
         return new_pose
 
-    def displce_pose(self,displacement_grasp_frame,pose):
+    def displace_pose(self,displacement_grasp_frame,pose):
         """
         Displaces a pose by the specified amount in the grasp frame.
 
         Args:
         displacement_grasp_frame (list): A list containing the x, y, z displacement in the grasp frame.
-        pose (PoseStamped): A ros PoseStamped message type to be displaces.
+        pose (Pose): A ros Pose message type to be displaces.
 
         Returns:
         PoseStamped: The displaced pose.
         """
-        rotation_matrix = self.quaternion_to_rotation_matrix([pose.pose.orientation.x, pose.pose.orientation.y, pose.pose.orientation.z, pose.pose.orientation.w])
+        rotation_matrix = self.quaternion_to_rotation_matrix([pose.orientation.x, pose.orientation.y, pose.orientation.z, pose.orientation.w])
         x_axis = rotation_matrix[:, 0]
         y_axis = rotation_matrix[:, 1]
         z_axis = rotation_matrix[:, 2]
-        new_pose = PoseStamped()
-        new_pose.header = pose.header
-        new_pose.pose.position.x = pose.pose.position.x + displacement_grasp_frame[0]*x_axis[0] + displacement_grasp_frame[1]*y_axis[0] + displacement_grasp_frame[2]*z_axis[0]
-        new_pose.pose.position.y = pose.pose.position.y + displacement_grasp_frame[0]*x_axis[1] + displacement_grasp_frame[1]*y_axis[1] + displacement_grasp_frame[2]*z_axis[1]
-        new_pose.pose.position.z = pose.pose.position.z + displacement_grasp_frame[0]*x_axis[2] + displacement_grasp_frame[1]*y_axis[2] + displacement_grasp_frame[2]*z_axis[2]
-        new_pose.pose.orientation = pose.pose.orientation
+        new_pose = Pose()
+        new_pose.position.x = pose.position.x + displacement_grasp_frame[0]*x_axis[0] + displacement_grasp_frame[1]*y_axis[0] + displacement_grasp_frame[2]*z_axis[0]
+        new_pose.position.y = pose.position.y + displacement_grasp_frame[0]*x_axis[1] + displacement_grasp_frame[1]*y_axis[1] + displacement_grasp_frame[2]*z_axis[1]
+        new_pose.position.z = pose.position.z + displacement_grasp_frame[0]*x_axis[2] + displacement_grasp_frame[1]*y_axis[2] + displacement_grasp_frame[2]*z_axis[2]
+        new_pose.orientation = pose.orientation
         return new_pose
 if __name__ == '__main__':
     try:
-        stretch_grasp_filter()
+        StretchGraspFilter()
     except rospy.ROSInterruptException:
         pass
     rospy.spin()
