--- stretch_fetch_grasp_bridge/srv/StretchGraspPose.srv
@@ -1,5 +1,0 @@
-int64 segment_no
-int64 rank_no
----
-bool success
-geometry_msgs/PoseStamped grasp_pose
