--- stretch_fetch_grasp_bridge/srv/StretchSegmentation.srv
@@ -1,4 +1,0 @@
-string object_name
----
-bool success
-sensor_msgs/PointCloud2 segmented_point_cloud
